# openapi/index.json
{"openapi":"3.1.0","info":{"title":"Nexus","license":{"name":"","identifier":"MIT"},"version":"0.1.0"},"paths":{"/users":{"get":{"tags":["Users"],"summary":"Get Users","operationId":"getUsers","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/User"},"type":"array","title":"Response Getusers"}}}},"418":{"description":"Error fetching user by ID"}}},"post":{"tags":["Users"],"summary":"Create User","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"409":{"description":"Conflict: Username already exists."},"418":{"description":"Error creating user"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{id}":{"get":{"tags":["Users"],"summary":"Get User By Id","operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"418":{"description":"Error fetching user by ID"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Users"],"summary":"Update User By Id","operationId":"updateUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"418":{"description":"Error patching user"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Users"],"summary":"Replace User By Id","operationId":"replaceUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"418":{"description":"Error updating user"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Users"],"summary":"Delete User By Id","operationId":"deleteUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDelete"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"418":{"description":"Error deleting user"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Audit":{"properties":{"created_at":{"type":"string","format":"date-time","title":"Created At"},"created_by":{"type":"integer","title":"Created By"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At","nullable":true},"updated_by":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Updated By","nullable":true},"archived_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Archived At","nullable":true},"archived_by":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Archived By","nullable":true},"unarchived_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Unarchived At","nullable":true},"unarchived_by":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Unarchived By","nullable":true},"is_archived":{"type":"boolean","title":"Is Archived"},"deleted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deleted At","nullable":true},"deleted_by":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Deleted By","nullable":true},"undeleted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Undeleted At","nullable":true},"undeleted_by":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Undeleted By","nullable":true},"is_deleted":{"type":"boolean","title":"Is Deleted"}},"type":"object","required":["created_at","created_by","is_archived","is_deleted"],"title":"Audit"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"User":{"properties":{"id":{"type":"integer","title":"Id"},"guid":{"type":"string","title":"Guid"},"username":{"type":"string","title":"Username"},"is_active":{"type":"boolean","title":"Is Active"},"audit":{"$ref":"#/components/schemas/Audit"}},"type":"object","required":["id","guid","username","is_active","audit"],"title":"User"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active","default":true}},"type":"object","required":["username","password"],"title":"UserCreate"},"UserDelete":{"properties":{"is_deleted":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Deleted"},"is_archived":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Archived"}},"type":"object","title":"UserDelete"},"UserUpdate":{"properties":{"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"},"password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"}},"type":"object","title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"Users","description":"User Management","endpoints":[{"operationId":"getUsers","summary":"Get Users"},{"operationId":"createUser","summary":"Create User"},{"operationId":"getUserById","summary":"Get User By Id"},{"operationId":"replaceUserById","summary":"Replace User By Id"},{"operationId":"updateUserById","summary":"Update User By Id"},{"operationId":"deleteUserById","summary":"Delete User By Id"}]}]}